refactor: Refactor into separate classes; various LoC reductions
