+++ Dockerfile
@@ -11,7 +11,7 @@ COPY --from=builder /install /usr/local
 
 WORKDIR /app
 ENV PYTHONUNBUFFERED=1
-COPY add_trackers.py .
+COPY *.py .
 
 # Run as nobody user
 USER nobody
