+++ app.py
@@ -1,0 +1,6 @@
+# telnetdoogie -  https://github.com/telnetdoogie/transmission-trackers
+from torrents_updater import TorrentUpdater
+
+if __name__ == "__main__":
+    app = TorrentUpdater()
+    app.main()
